Refactor and lint proxies/api_umbrella directory
