+++ proxies/api_umbrella/collections/index.js
@@ -1,0 +1,6 @@
+import { Mongo } from 'meteor/mongo';
+
+const ApiUmbrellaAdmins = new Mongo.Collection('apiUmbrellaAdmins');
+const ApiUmbrellaUsers = new Mongo.Collection('apiUmbrellaUsers');
+
+export { ApiUmbrellaAdmins, ApiUmbrellaUsers };
