--- proxies/api_umbrella/collections/server/api_umbrella_admins_publications.js
@@ -1,3 +1,0 @@
-Meteor.publish('apiUmbrellaAdmins', function () {
-  return ApiUmbrellaAdmins.find();
-});
