--- proxies/api_umbrella/collections/server/api_umbrella_users_publications.js
@@ -1,3 +1,0 @@
-Meteor.publish('apiUmbrellaUsers', function () {
-  return ApiUmbrellaUsers.find();
-});
