+++ proxies/api_umbrella/collections/server/publications.js
@@ -1,0 +1,13 @@
+import { Meteor } from 'meteor/meteor';
+
+import { ApiUmbrellaAdmins, ApiUmbrellaUsers } from '../';
+
+Meteor.publish('apiUmbrellaAdmins', () => {
+  // Get ApiUmbrellaAdmins collection object
+  return ApiUmbrellaAdmins.find();
+});
+
+Meteor.publish('apiUmbrellaUsers', () => {
+  // Get ApiUmbrellaUsers collection object
+  return ApiUmbrellaUsers.find();
+});
